LibWeb: Split GridTrackPlacementShorthandStyleValue out of StyleValue
